Split Avl in two: immutable Avl and MvccAvl
Before this changeset, Avl was implemented with interior mutability, the
immutable interface was an implementation detail exposed internally
through `Node` type API. Behind the scenes, snapshots of the tree were
used for reading operations, essentially implementing multi-version
concurrency control in its most primitive form.

Since snapshots were hidden from the interface, we had to resort to
`unsafe` to reassign the lifetimes in the `iter` method. There are other
uses for the immutable version of the AVL that I'd like to
explore (specifically, the index entry list container).

Now, `Avl` exposes the immutable API (all modifying operations return a
new tree instead of changing the current one) and in addition a new type
`MvccAvl` is introduced that preserves the old interface and behaviour
of `Avl`.