Use AVL instead of List to store index entries
For bigger indices, when the removing from the index is performed and
the list of entries for the given word is long enough, the programme
will crash with stack overflow.

This happens due to recursion in `List`'s destructor, the head is dropped,
then, recursively, the rest of the tail is dropped. To lower the depth
of recursion on drop, we now use `Avl` to store entries.