Improve performance of removing from an index
Now that we use `Avl` to store index entries, it is trivial to add
grouping of entries by path. This allows log(n) removes of an entry from
the list of entries for the word.

As another optimisation, we now store the set of words that were read
from the original file, to not walk the whole tree for every removed file.